Consolidate integration tests
The Conan package named `beman` provides package tests that
test whether a downstream package can depend on beman libraries
and use them together. Instead of using a specific test suite
for this purpose, this commit alters the package tests for the
`beman` package to use the tests defined at the top-level of
this repo.

This allows the same test suite to support both the `FetchContent`
and the Conan dependency management workflows.