Add rules that mitigate ODR violations
The three rules added here are generally practiced in Beman repositories
as a means to mitigate ODR violation risks. While they have been discussed
at length in various PRs and issues (see
[this comment](https://github.com/bemanproject/exemplar/issues/40#issuecomment-2388995030)
for an example), this repeatedly comes up with new folks joining the
project.

By encoding this into rules linked to a PR pointing to the rationale
the hope is to gain consistency and avoid rehashing the discussions.